requirements.txt
pyautogui==0.9.54
keyboard==0.13.5
PyQt5==5.15.9
Pillow==10.2.0 